Add: Tony's work to the final project redesign by me
All the source code that Tony wrote, I looked at it and transfert his
work in the final projet. I changed a lot of "id" names and the
structure a little bit to respect the W3C rules. I also had an idea
about the "About" and "Contact" page. I changed it to a box that will be
displayed when we click on the link of the nav bar.